feature #28: refactor home screen
