+++ feature/src/main/java/com/eoyeongbooyeong/home/HomeSideEffect.kt
@@ -1,0 +1,5 @@
+package com.eoyeongbooyeong.home
+
+sealed class HomeSideEffect {
+    data class NavigateToWebView(val url: String) : HomeSideEffect()
+}
