+++ feature/src/main/java/com/eoyeongbooyeong/home/HomeState.kt
@@ -1,0 +1,5 @@
+package com.eoyeongbooyeong.home
+
+data class HomeState(
+    val temp: String = ""
+)
