--- feature/src/main/java/com/eoyeongbooyeong/new_home/NewHomeSideEffect.kt
@@ -1,5 +1,0 @@
-package com.eoyeongbooyeong.new_home
-
-sealed class NewHomeSideEffect {
-    data class NavigateToWebView(val url: String) : NewHomeSideEffect()
-}
