--- feature/src/main/java/com/eoyeongbooyeong/new_home/NewHomeState.kt
@@ -1,5 +1,0 @@
-package com.eoyeongbooyeong.new_home
-
-data class NewHomeState(
-    val temp: String = ""
-)
